+++ Magic/CHANGELOG.md
@@ -15,7 +15,8 @@
    - (WIP, will be tameable with items)
    - TODO: add pets example with pet shop, ownable pets
  - Add `collidable` flag to mobs
- - Add `stay`, `owner` and `owned` flags to `CheckEntity`, for specific behaviors for owned mobs  
+ - Add `stay`, `owner` and `owned` flags to `CheckEntity`, for specific behaviors for owned mobs
+ - Add `owned` and `stay` requirements  
  - New `moon` requirement for phase-of-moon logic, simplify `time` requirement   
  - Example `scaredy` mob added, a cat that gets scared in the dark
  - Example `philbert` mob added, a friendly tameable creeper